[MISC] Avoid circular dependencies
