+++ pymonad/promise.py
@@ -93,9 +93,9 @@ accepts an Exception as it's input.
 
 This program prints "<type 'IndexError'>" as its output.
 """
-from typing import Callable, Generic, TypeVar, Union
-
 import asyncio
+from typing import Callable, Generic, TypeVar, Union, Awaitable
+
 import pymonad.monad
 import pymonad.tools
 
@@ -227,6 +227,68 @@ def Promise(function: PromiseFunction) -> _Promise[T]: # pylint: disable=invalid
     async def _awaitable(function, resolve, reject):
         return function(resolve, reject)
     return _Promise(_awaitable(function), None) # pylint: disable=no-value-for-parameter
+
+
+def async_func(func: Callable) -> Callable:
+    """Transform simple function in async function using promises.
+
+    async_func is supposed to be used as a decorator for providing
+    seamless integration with asynchronous operations, by deferring
+    arguments and keyword arguments input to be scheduled asynchronously
+    into the event-loop. It also transform the output into a promise,
+    allowing for abstract computation.
+
+    Example:
+      @async_func
+      def add(x, y):
+          return x+y
+
+      x = (Promise.insert(1)
+                      .then(long_id))
+      y = (Promise
+              .insert(2)
+              .then(long_id)
+              .then(div(0))            # Raises an error...
+              .catch(lambda error: 2)) # ...which is dealth with here.
+
+      z = add(x, y)
+
+      # z is now a Promise object that can be further used into other
+      # operations, abstracting chain of computations
+
+      print( await z.map(long_id).catch(lambda error: 'Recovering...') )
+
+    Args:
+      func: a regular function with variable arguments args and keyword
+            arguments kwargs
+
+     Returns:
+      An asynchronous counterpart of the provided function, that
+      accepts both promise and/or regular values and  returns promises,
+      for asynchronous usage
+
+    """
+
+    async def getArgs(args):
+        return await asyncio.gather(
+            *[arg if isinstance(arg, Awaitable) else Promise.insert(arg) for arg in args]
+        )
+
+    async def getKwargs(kwargs):
+        kwargsTasks = [arg.map(lambda x: (ith, x)) if isinstance(arg, Awaitable) else Promise.insert((ith, arg))
+                       for ith, arg in kwargs.items()]
+        return dict(await asyncio.gather(*kwargsTasks))
+
+    async def async_wrap(*args, **kwargs):
+        (_args, _kwargs) = await asyncio.gather(getArgs(args), getKwargs(kwargs))
+
+        return func(*_args, **_kwargs)
+
+    def wrapper(*args, **kwargs):
+        return _Promise(lambda resolve, reject: async_wrap(*args, **kwargs), None)
+
+    return wrapper
+
 
 Promise.apply = _Promise.apply
 Promise.insert = _Promise.insert
